Bug fix
Smoothed over the bug, so the jerky movement should be gone. removed the code to smooth the encounterobjects to it's own script, for safety's sake.

The rocket seems to be a square now. I'll have to fix that later.

the glitch saying the level can't end is gone, but there doesn't appear to be level switching. could whoever is managing that tell what i'm doing wrong?